requirements.txt
flask
flask-wtf
wtforms
python-docx
azure-storage-blob
pdf2docx
